requirements.txt
asgiref==3.8.1
Django==5.1.5
django-taggit==6.1.0
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.1
